+++ cart-service/src/main/java/com/atozmart/cart/dto/ViewCartResponse.java
@@ -1,11 +1,12 @@
 package com.atozmart.cart.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ViewCartResponse(
 
 		List<ItemDto> items,
 
-		double orderAmount) {
+		BigDecimal orderAmount) {
 
 }
