+++ cart-service/src/main/java/com/atozmart/cart/dto/order/OrderItemsDto.java
@@ -1,15 +1,17 @@
 package com.atozmart.cart.dto.order;
 
+import java.math.BigDecimal;
+
 public record OrderItemsDto(
 
 		String itemId,
 
 		String itemName,
 
-		double unitPrice,
+		BigDecimal unitPrice,
 
 		int quantity,
 
-		double effectivePrice) {
+		BigDecimal effectivePrice) {
 
 }
